% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-summaries.R
\name{assign_neighborhoods}
\alias{assign_neighborhoods}
\title{Assign bins to fitted spatial neighborhoods}
\usage{
assign_neighborhoods(object, grids, append = TRUE)
}
\arguments{
\item{object}{A fitted object from [fit_spatial_neighborhoods()].}

\item{grids}{Grid table to assign.}

\item{append}{Return the grid with a neighborhood column when TRUE.}
}
\value{
Integer assignments or the augmented grid.
}
\description{
Assign bins to fitted spatial neighborhoods
}
