% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-summaries.R
\name{patient_neighborhood_prevalence}
\alias{patient_neighborhood_prevalence}
\title{Patient-level neighborhood prevalence}
\usage{
patient_neighborhood_prevalence(grids)
}
\arguments{
\item{grids}{Assigned grid returned by [assign_neighborhoods()].}
}
\value{
Long-form patient-level prevalence table.
}
\description{
Patient-level neighborhood prevalence
}
